migrate all components to C23 standard
- config.mk: -std=c23 across all 7 components (slock, slstatus, imgdec,
  dmenu, netpanel, dwm core, st)
- NULL → nullptr (~627 spots)
- [[noreturn]] on die() in all util.h/util.c
- [[fallthrough]] replacing comment-based fallthroughs
- [[maybe_unused]] on ~35+ unused params (Clang strict mode)
- Sign-compare fixes (~50+): proper casts for size_t/unsigned int
- Format fixes: %zu for size_t, snprintf, (unsigned char) subscripts
- Vendored headers (khash.h, kvec.h, stb/*.h) untouched, warnings
  suppressed with #pragma GCC diagnostic push/pop
- CC = gcc in st/config.mk (c99 doesn't support -std=c23)
- netpanel/netpanel binary removed from tracking, added to .gitignore

Verified: 4-layer matrix (GCC strict, Clang strict, ASan/UBSan,
functional smoke test) — 0 errors across all 7 components.

+++ shiftview.c
@@ -12,13 +12,13 @@ void
 shiftview(const Arg *arg) {
   Arg shifted;
 
-  if(arg->i > 0 && arg->i < LENGTH(tags)) // left circular shift
+  if(arg->i > 0 && arg->i < (int)LENGTH(tags)) // left circular shift
     shifted.ui = (selmon->tagset[selmon->seltags] << arg->i)
-      | (selmon->tagset[selmon->seltags] >> (LENGTH(tags) - arg->i));
+      | (selmon->tagset[selmon->seltags] >> ((int)LENGTH(tags) - arg->i));
 
-  else if(arg->i < 0 && -arg->i < LENGTH(tags)) // right circular shift
+  else if(arg->i < 0 && -arg->i < (int)LENGTH(tags)) // right circular shift
     shifted.ui = selmon->tagset[selmon->seltags] >> (- arg->i)
-      | selmon->tagset[selmon->seltags] << (LENGTH(tags) + arg->i);
+      | selmon->tagset[selmon->seltags] << ((int)LENGTH(tags) + arg->i);
 
   else /* out-of-range shift: skip to avoid undefined behavior / garbage tags */
     return;
